requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.1
langchain-core>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
pandas>=2.0.0
scikit-learn>=1.0.0

